Extract detail view + commandcenter wiring
